Implemented the cache. Defined node IDs on 7 bits. Moved Node::send(payload,nodeID,network) to Messaging where it belongs.

+++ Core/mdaemon_node.cpp
@@ -35,7 +35,7 @@ namespace	mBrane{
 	namespace	sdk{
 		namespace	mdaemon{
 
-			Node::Node(uint16	ID):module::Node(ID),_shutdown(false){
+			Node::Node(uint8	ID):module::Node(ID),_shutdown(false){
 			}
 
 			Node::~Node(){
