+++ Core/mdaemon_node.h
@@ -58,7 +58,7 @@ namespace	mBrane{
 				StaticArray<DynamicClassLoader<Daemon>	*>	daemonLoaders;
 				StaticArray<Daemon	*>						daemons;
 				StaticArray<Thread	*>						daemonThreads;
-				Node(uint16	ID=NoID);
+				Node(uint8	ID=NoID);
 				bool	loadConfig(XMLNode	&n);
 				void	start();
 				virtual	void	shutdown();
@@ -67,9 +67,9 @@ namespace	mBrane{
 				bool	isRunning();
 				virtual	void				dump(const	char	*fileName)=0;	//	dumps the current system state; module dump fileNames: module_class_ID.bin: ex: CR1_123.bin
 				virtual	void				load(const	char	*fileName)=0;	//	initializes itself from a previously saved system state
-				virtual	void				migrate(uint16	CID,uint16	ID,uint16	NID)=0;	//	identifies the module and the target node
+				virtual	void				migrate(uint16	CID,uint16	ID,uint8	NID)=0;	//	identifies the module and the target node
 				//virtual	Array<uint8,65535>	&sharedMemorySegment(uint8	segment)=0;	//	FUTURE DEVELOPMENT: pinned down RAM for RDMA, 16KB-1
-				virtual	module::_Module		*getModule(uint16	hostID,uint16	CID,uint16	ID)=0;	//	so that the daemons can write module internals and use modules as interfaces to the pub-sub network
+				virtual	module::_Module		*getModule(uint8	hostID,uint16	CID,uint16	ID)=0;	//	so that the daemons can write module internals and use modules as interfaces to the pub-sub network
 				//	TODO:	define API as pure virtual functions
 				//			-> node map (an array of mBrane::Networking::NetworkID)
 				//			-> profiling data
